+++ OUDS/Core/Components/Sources/CheckboxPicker/OUDSCheckboxPicker.swift
@@ -315,7 +315,7 @@ public struct OUDSCheckboxPicker<Tag>: View where Tag: Hashable {
     /// Forges the label to vocalize with *Voice Over* for the root element
     private var a11yLabel: String {
         if case let .verticalRooted(label, _) = placement {
-            label
+            label + (isError ? ", " + "core_common_onError_a11y".localized() : "")
         } else {
             ""
         }